fix: split if tests into 2
