+++ app/components/atoms/If/tests/index.test.js
@@ -9,7 +9,6 @@
 import React from 'react';
 import { Text } from 'react-native';
 import { render } from '@testing-library/react-native';
-import set from 'lodash/set';
 import If from '../index';
 
 describe('<If />', () => {
@@ -17,8 +16,9 @@ describe('<If />', () => {
     const baseElement = render(<If />);
     expect(baseElement).toMatchSnapshot();
   });
-  it('Should render the correct prop component based on the condition', () => {
-    const conditionTrueText = 'Shoudld render when condition is true';
+
+  it('If should render the child component when props.condition = true', () => {
+    const conditionTrueText = 'Should render when condition is true';
     const conditionFalseText = 'Should render condition is false';
     const OtherwiseComponent = () => <Text>{conditionFalseText}</Text>;
     const TrueConditionComponent = () => <Text>{conditionTrueText}</Text>;
@@ -26,18 +26,31 @@ describe('<If />', () => {
       otherwise: <OtherwiseComponent />,
       condition: true
     };
-    const { getByText } = render(
+    const { queryByText, getByText } = render(
       <If {...props}>
         <TrueConditionComponent />
       </If>
     );
     expect(getByText(conditionTrueText)).toBeTruthy();
-    set(props, 'condition', false);
-    const { getByText: textQueryOnReRender } = render(
+    expect(queryByText(conditionFalseText)).not.toBeTruthy();
+  });
+
+  it('If should render the component passed on otherwise when props.condition = false', () => {
+    const conditionFalseText = 'Should render condition is false';
+    const conditionTrueText = 'Should render when condition is true';
+    const TrueConditionComponent = () => <Text>{conditionTrueText}</Text>;
+    const OtherwiseComponent = () => <Text>{conditionFalseText}</Text>;
+    const props = {
+      otherwise: <OtherwiseComponent />,
+      condition: false
+    };
+
+    const { queryByText, getByText } = render(
       <If {...props}>
         <TrueConditionComponent />
       </If>
     );
-    expect(textQueryOnReRender(conditionFalseText)).toBeTruthy();
+    expect(getByText(conditionFalseText)).toBeTruthy();
+    expect(queryByText(conditionTrueText)).not.toBeTruthy();
   });
 });
